refactor: migrate to solid_lints 1.0.0-dev.1 plugin and fix lints refactor: reduce cyclomatic complexity of some methods refactor: reduce numbers of parameters by extracting a class docs: update changelog
