+++ analysis_options.yaml
@@ -1,1 +1,14 @@
 include: package:solid_lints/analysis_options.yaml
+
+plugins:
+  solid_lints: 1.0.0-dev.1
+
+solid_lints:
+  diagnostics:
+    named_parameters_ordering:
+      order:
+        - required
+        - required_super
+        - default
+        - nullable
+        - super
