+++ example/analysis_options.yaml
@@ -1,6 +1,9 @@
 include: package:solid_lints/analysis_options.yaml
 
-custom_lint:
-  rules:
-    - avoid_late_keyword:
+plugins:
+  solid_lints: 1.0.0-dev.1
+
+solid_lints:
+  diagnostics:
+    avoid_late_keyword:
       allow_initialized: true
