+++ example/pubspec.yaml
@@ -16,7 +16,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  solid_lints: ^0.3.1
+  solid_lints: 1.0.0-dev.1
 
 flutter:
   uses-material-design: true
