+++ lib/languagetool_textfield.dart
@@ -6,11 +6,14 @@ export 'src/core/enums/delay_type.dart';
 export 'src/core/model/category.dart';
 export 'src/core/model/language_tool_raw.dart';
 export 'src/core/model/match.dart';
+export 'src/core/model/match_location.dart';
 export 'src/core/model/replacement.dart';
 export 'src/core/model/rule.dart';
 export 'src/domain/highlight_style.dart';
 export 'src/domain/language_check_service.dart';
 export 'src/domain/mistake.dart';
+export 'src/domain/mistake_colors.dart';
+export 'src/domain/mistake_popup_style.dart';
 export 'src/domain/writing_mistake.dart';
 export 'src/language_check_services/language_tool_service.dart';
 export 'src/presentation/language_tool_text_field.dart';
