+++ lib/src/domain/highlight_style.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:languagetool_textfield/src/domain/mistake_colors.dart';
 
 /// Class creates color scheme for highlighting mistakes
 class HighlightStyle {
@@ -6,26 +7,8 @@ class HighlightStyle {
   static const double _initialBackgroundOpacity = 0.2;
   static const double _initialLineHeight = 1.5;
 
-  /// Misspelling mistake highlight color
-  final Color misspellingMistakeColor;
-
-  /// Misspelling mistake highlight color
-  final Color typographicalMistakeColor;
-
-  /// Typographical mistake highlight color
-  final Color grammarMistakeColor;
-
-  /// Uncategorized mistake highlight color
-  final Color uncategorizedMistakeColor;
-
-  /// NonConformance mistake highlight color
-  final Color nonConformanceMistakeColor;
-
-  /// Style mistake highlight color
-  final Color styleMistakeColor;
-
-  /// Other mistake highlight color
-  final Color otherMistakeColor;
+  /// Highlight color of each mistake type
+  final MistakeColors colors;
 
   /// background opacity for mistake TextSpan
   final double backgroundOpacity;
@@ -36,15 +19,9 @@ class HighlightStyle {
   /// Mistaken text decoration style
   final TextDecoration decoration;
 
-  ///Color scheme constructor
+  /// Color scheme constructor
   const HighlightStyle({
-    this.misspellingMistakeColor = Colors.red,
-    this.typographicalMistakeColor = Colors.green,
-    this.grammarMistakeColor = Colors.amber,
-    this.uncategorizedMistakeColor = Colors.blue,
-    this.nonConformanceMistakeColor = Colors.greenAccent,
-    this.styleMistakeColor = Colors.deepPurpleAccent,
-    this.otherMistakeColor = Colors.white60,
+    this.colors = const MistakeColors(),
     this.backgroundOpacity = _initialBackgroundOpacity,
     this.mistakeLineThickness = _initialLineHeight,
     this.decoration = TextDecoration.underline,
